requirements.txt
Django==2.0.2
pytz>=2018.3
setuptools>=38.5.1
django-simple-search>=1.0.2